[Isolates] Allow running multiple isolates in shell and use this in tests.
Sample usage (prints a sequence of mixed 1's and 2's):
  $ ./shell_g -e 'while (1) print(1)' --isolate -e 'while (1) print(2)'

The first group of source files (before the first --isolate) is run in
the default isolate. All the other groups are run in separate isolate
threads.

It would be nice to use our platform.h but it drags too many includes,
so I'm using pthreads on anything non-WIN32 for now.

Review URL: http://codereview.chromium.org/3601010

git-svn-id: https://v8.googlecode.com/svn/branches/experimental/isolates@5599 ce2b1a6d-e550-0410-aec6-3dcde31c8c00