[Isolates] Fix auto extraction of isolate from heap objects in handle constructor.
My previous attempt used Handle(T*) and Handle(HeapObject*) overloads,
but since T is a class template parameter (and not just a paramer of
the constructor) the first signature is preferred when the argument is
convertible to T*.

This patch exposed an issue with SaveContext which created a handle
pointing to NULL context.

Bonus: a few lint fixes.

Review URL: http://codereview.chromium.org/3788008

git-svn-id: https://v8.googlecode.com/svn/branches/experimental/isolates@5662 ce2b1a6d-e550-0410-aec6-3dcde31c8c00